Rollback Magick.NET upgrade

+++ PhotoManager/PhotoManager.Common/BitmapHelper.cs
@@ -118,7 +118,7 @@ public static class BitmapHelper
                     MagickImageApplyRotation(magickImage, rotation, false); // Apply Rotation because MagickImage does not rotate the image in-place
 
                     // Resize the MagickImage
-                    magickImage.Resize((uint)width, (uint)height);
+                    magickImage.Resize(width, height);
 
                     // Convert the MagickImage to a BitmapImage
                     using (MemoryStream bitmapStream = new())
